% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotMethods.R
\name{buildSpotSpace}
\alias{buildSpotSpace}
\title{Constructor of PathwaySpace-class Objects}
\usage{
buildSpotSpace(
  spot_coord,
  raster_image,
  mar = 0.1,
  nrc = 500,
  flip.x = FALSE,
  flip.y = TRUE,
  rotate.xy = FALSE,
  crop_coord = c(0, 1, 0, 1),
  verbose = TRUE
)
}
\arguments{
\item{spot_coord}{A data frame with spot coordinates. It must include 
row names and coordinates assigned to `x` and `y` column names.}

\item{raster_image}{A raster image serving as the background on which 
spot coordinates are mapped.}

\item{mar}{A single numeric value (in [0,1]) indicating the size of the 
outer margins as a fraction of the image matrix defined by `nrc`.}

\item{nrc}{A single positive integer indicating the number of rows and 
columns of a square image matrix onto which spot signals will be projected.}

\item{flip.x}{Logical; whether to flip the `raster_image` along the x-axis.}

\item{flip.y}{Logical; whether to flip the `raster_image` along the y-axis.}

\item{rotate.xy}{Logical; whether to rotate the `raster_image`.}

\item{crop_coord}{An optional numeric vector of length four specifying a
cropping region (xmin, xmax, ymin, ymax), with values in normalized 
coordinates [0,1].}

\item{verbose}{Logical; whether to display detailed messages.}
}
\description{
\code{buildSpotSpace} constructs a PathwaySpace-class object 
by integrating spatial spot coordinates with a raster image background.
}
\details{
Proper spatial alignment between spot coordinates and the background image 
depends on consistent coordinate conventions. Different graphics systems 
and external software may adopt distinct origins and rendering directions 
(e.g., top-left vs. bottom-left). To address these differences, 
\code{buildSpotSpace} provides orientation control through \code{rotate.xy}, 
\code{flip.x}, and \code{flip.y}. If misalignment is observed, users 
should test alternative combinations of these arguments to achieve 
correct spatial alignment.
}
\examples{
# See examples in the SpotSpace's vignette:
# vignette("SpotSpace")

}
\seealso{
\code{\link[PathwaySpace]{buildPathwaySpace}}
}
\author{
Sysbiolab Team
}
