% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotMethods.R
\name{getNearestNode}
\alias{getNearestNode}
\title{getNearestNode}
\usage{
getNearestNode(ps)
}
\arguments{
\item{ps}{A \code{\link[PathwaySpace]{PathwaySpace}} class object.}
}
\description{
\code{getNearestNode} retrieves the nearest neighbor for 
each spot based on Euclidean distances.
}
\examples{
# See examples in the SpotSpace's vignette:
vignette("SpotSpace")

}
\seealso{
\code{\link[RANN]{nn2}}
}
