% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotMethods.R
\name{getNearestSpot}
\alias{getNearestSpot}
\title{getNearestSpot}
\usage{
getNearestSpot(ps)
}
\arguments{
\item{ps}{A \code{\link[PathwaySpace]{PathwaySpace}} class object.}
}
\description{
\code{getNearestSpot} retrieves the nearest neighbor for 
each spot based on Euclidean distances.
}
\examples{
# See examples in the SpotSpace's vignette:
vignette("SpotSpace")

}
\seealso{
\code{\link[RANN]{nn2}}
}
