% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotMethods.R
\name{buildSpotSpace}
\alias{buildSpotSpace}
\title{Constructor for GraphSpace-class objects from spot coordinates}
\usage{
buildSpotSpace(
  spot_coord,
  raster_image,
  mar = 0.1,
  flip.v = FALSE,
  flip.h = FALSE,
  flip.x = FALSE,
  flip.y = TRUE,
  rotate.xy = FALSE,
  verbose = TRUE,
  crop_coord = deprecated(),
  nrc = deprecated()
)
}
\arguments{
\item{spot_coord}{A data frame with spot coordinates. It must include 
row names and coordinates assigned to `x` and `y` column names.}

\item{raster_image}{A raster image serving as the background on which 
spot coordinates are mapped.}

\item{mar}{A single numeric value in \code{[0, 0.5]} for the graph margins.}

\item{flip.v}{Logical; whether to vertically flip the background image
matrix (top-to-bottom) to align with the graph coordinate system.}

\item{flip.h}{Logical; whether to horizontally flip the background image
matrix (left-to-right) to align with the graph coordinate system.}

\item{flip.x}{Logical; whether to flip the node coordinates along the x-axis.}

\item{flip.y}{Logical; whether to flip the node coordinates along the y-axis.}

\item{rotate.xy}{Logical; whether to rotate graph coordinates.}

\item{verbose}{Logical; whether to display detailed messages.}

\item{crop_coord}{`r lifecycle::badge("deprecated")` Deprecated since
SpotSpace 0.0.7; use \link[RGraphSpace]{cropGraphSpace} instead.}

\item{nrc}{`r lifecycle::badge("deprecated")` Deprecated since
SpotSpace 0.0.7; use \link[PathwaySpace]{buildPathwaySpace} instead.}
}
\value{
A \code{\link[RGraphSpace:GraphSpace-class]{GraphSpace}} object with 
integrated spatial and image data.
}
\description{
\code{buildSpotSpace} constructs a \code{GraphSpace-class} object 
by integrating spatial spot coordinates with a raster background. This 
wrapper streamlines data preparation for graph and spatial workflows, 
such as in the \code{PathwaySpace} package.
}
\details{
Spatial alignment between spot coordinates and background images 
often requires orientation adjustments due to differing coordinate 
conventions (e.g., top-left vs. bottom-left origins). \code{buildSpotSpace} 
automates the initial object creation and calls \code{normalizeGraphSpace} 
internally to handle these adjustments via \code{rotate.xy}, \code{flip.x}, 
and \code{flip.y}.
}
\examples{
# See examples in the PathwaySpace's online tutorials:
# https://sysbiolab.github.io/PathwaySpace/

}
\seealso{
\link[RGraphSpace]{GraphSpace}, 
\link[RGraphSpace]{normalizeGraphSpace}, 
\link[PathwaySpace]{buildPathwaySpace}
}
\author{
Sysbiolab Team
}
