% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotMethods.R
\name{buildSpotSpace}
\alias{buildSpotSpace}
\title{Constructor of PathwaySpace-class Objects}
\usage{
buildSpotSpace(
  spot_coord,
  raster_image,
  mar = 0.1,
  nrc = 500,
  rotate.xy = TRUE,
  flip.y = TRUE,
  flip.x = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spot_coord}{A data frame with spot coordinates. It must include 
row names and coordinates assigned to `x` and `y` column names.}

\item{raster_image}{A raster image serving as the background on which 
spot coordinates are mapped.}

\item{mar}{A single numeric value (in [0,1]) indicating the size of the 
outer margins as a fraction of the image matrix defined by `nrc`.}

\item{nrc}{A single positive integer indicating the number of rows and 
columns of a square image matrix onto which spot signals will be projected.}

\item{rotate.xy}{Logical; whether to rotate the `raster_image`.}

\item{flip.y}{Logical; whether to flip the `raster_image` along the y-axis.}

\item{flip.x}{Logical; whether to flip the `raster_image` along the x-axis.}

\item{verbose}{Logical; whether to display detailed messages.}
}
\description{
\code{buildSpotSpace} is a constructor of
PathwaySpace-class objects.
}
\examples{
# Load a demo igraph
data('gtoy1', package = 'RGraphSpace')

# Check graph validity
gs <- GraphSpace(gtoy1, mar = 0.1)

# Create a new PathwaySpace object

}
\seealso{
\code{\link[PathwaySpace]{buildPathwaySpace}}
}
\author{
Sysbiolab Team
}
